Run the catalogue against SQLite, and let the write queue do the locking
The catalogue is the package everything else was waiting behind. It goes
from 266 failing cases to two, and `packages/recipes` is green as a result.

**Two constructs are gone and nothing replaces them, for one reason.** Two
advisory locks and eleven row-locking clauses were all arranging the same
thing: that a second writer could not slip between a read and the write it
justified. Every one of those paths now runs inside a transaction the venue
file's write queue serialises -- one writer, for the whole file, which is
wider than any of them. So they are deleted rather than translated.

That is a claim, so it has a receipt. A fixture helper starts two write
transactions together, holds the first open, and reports whether the second
ran a statement of its own: through the queue it reports no, and with the
same two bodies started outside the queue it reports yes. Swapping the
helper's second body to the unqueued form makes the two converted
serialisation cases fail, and restoring it makes them pass. Every converted
concurrency case in the package goes through that helper, so they all carry
the observation rather than resting on the argument.

Four helpers are renamed to what they now do -- a function called "lock"
that takes no lock is a false claim -- and the whole pattern is written out
once, on the extras one, with its measurement and its control. The other
forty-odd row-locking sites in the tree can follow it.

Eight failure classes, not the two that were known. Casts and array
constructors; PostgreSQL catalogue views replaced by the engine's own table
and index pragmas; a raw insert that omitted a column whose generator moved
into JavaScript; a read of a JSON column that hands back the stored text
rather than a parsed value; and error matching, where this driver reports
the same code for every failure alike and only a second field discriminates.

What the engine will not tell us cost several assertions outright: a
foreign-key refusal's whole message is that one failed, with no constraint
name and no column. Those assertions are gone with no replacement. Checks
keep their names, so those survive word for word, and unique violations are
matched by which table and columns collided, which is the question the index
name used to answer.

The schema suite was rewritten against the engine's pragmas, and the
rewrite was not taken from the new database and left there: all sixty of the
old PostgreSQL definitions -- 23 foreign keys, 17 primary keys, 5 unique
constraints, 14 checks, 13 lookup indexes -- were checked one at a time
against the SQLite facts. Every one carries. Two changed shape and not
meaning: foreign and primary keys have no name here, so the map is keyed by
table and columns, and one check lost the null half of its subject, which an
earlier commit on this branch had already given up in the schema itself.

Eight grant and role walkthroughs are deleted, because there are no grants
and no roles to walk. Two more are deleted because the property they proved
is now false by design: two saves of one list DO wait for each other, and a
transaction cannot be interrupted between its reads. Each deletion is
listed with what it proved and where the behaviour survives. One is a real
loss with no replacement anywhere -- a reassignment that used to be refused
unless a table's primary key doubled as its replication identity, which is
not a thing this engine has -- and the comment that named that test is
corrected. One case that looked unreachable was restaged rather than dropped:
its torn read cannot happen, but the state it produced can be written
directly, and the control confirms the case still fails without it.

Two cases are left FAILING rather than edited to pass, and one of them is a
product regression this change surfaces rather than causes. A product id
sent in upper case used to be settled by the database, because the column
was a type that normalises its input; it is ordinary text now and compares
byte for byte, so the row is refused. Measured with a control in both
directions. How ids get normalised is a decision wider than this package,
so the assertion stands as it was, with a note. The other is a rounding
control that compared a SQL cast against our own converter: the two agreed
on PostgreSQL and no longer do, because SQLite has no exact decimal, so
changing the expected value would assert the opposite of what the line is
for. Both readings, and a control in the other direction, are in the note.

+++ packages/catalogue/src/operations.test.ts
@@ -55,7 +55,7 @@ import type { AvailableProduct } from "./operations.js";
 import { createUnit, EACH_UNIT, readProductUnitId } from "./units.js";
 import { seedCatalogueFixture, seedVenue, useCatalogueDb } from "../test/fixtures.js";
 
-// Query behaviour runs on PGlite; each case starts with empty authoring tables.
+// Query behaviour. Each case starts with empty authoring tables.
 const fx = useCatalogueDb();
 
 describe("catalogue operations", () => {
@@ -219,7 +219,7 @@ describe("catalogue operations", () => {
     });
   });
 
-  // Every test body runs as app_user on its own transaction.
+  // Every test body runs on its own transaction.
   const asTenant = <T>(fn: (tx: Transaction) => Promise<T>): Promise<T> =>
     withTransaction(fx.db, async (tx) => {
       await asAppUser(tx);
@@ -1381,7 +1381,7 @@ describe("catalogue operations", () => {
         { id: casa.id, name: "Casa", isDefault: true },
       ]);
       const members = await tx.execute<{ count: number }>(
-        sql`select count(*)::int as count from location_catalogues where location_id = ${locationId}`,
+        sql`select count(*) as count from location_catalogues where location_id = ${locationId}`,
       );
       expect(members.rows[0]!.count).toBe(0);
     });
